+++ CHANGELOG.md
@@ -1,11 +1,18 @@
 # Changelog
 
-## [Unreleased](https://github.com/schireson/pytest-alembic/compare/v0.11.0...HEAD) (2025-05-16)
+### [v0.12.1](https://github.com/schireson/pytest-alembic/compare/v0.12.0...v0.12.1) (2025-05-22)
+
+#### Fixes
+
+* Handle pyproject.toml based alembic config.
+([82d9d62](https://github.com/schireson/pytest-alembic/commit/82d9d62024ccb86b655272a70e0d94374b5675a5))
+
+## [v0.12.0](https://github.com/schireson/pytest-alembic/compare/v0.11.1...v0.12.0) (2025-05-16)
 
 ### Fixes
 
 * Release 0.12.0
-([87dbf29](https://github.com/schireson/pytest-alembic/commit/87dbf29e24f47417a5821fb6908673058ead7d89))
+([4baee1e](https://github.com/schireson/pytest-alembic/commit/4baee1e85abe3733e6aa785c7704248f4fbb59b0))
 * Failing tests.
 ([1dc5b43](https://github.com/schireson/pytest-alembic/commit/1dc5b43d7e5b7530181808490cb31386631990a9))
 * Linting.
@@ -16,6 +23,13 @@
 ([853b116](https://github.com/schireson/pytest-alembic/commit/853b1164af9d03cb71504895c63e4962cc14ab84))
 * Updated pyproject.toml to be more flexible with poetry_core versioning.
 ([c7f25c3](https://github.com/schireson/pytest-alembic/commit/c7f25c39a795591a794f1ab89203f4a992012319))
+
+### [v0.11.1](https://github.com/schireson/pytest-alembic/compare/v0.11.0...v0.11.1) (2024-03-27)
+
+#### Fixes
+
+* Ensure branched revisions are upgraded individually once.
+([25b03a3](https://github.com/schireson/pytest-alembic/commit/25b03a3cac04258bbf3725f00e4794663808581a))
 
 ## [v0.11.0](https://github.com/schireson/pytest-alembic/compare/v0.10.7...v0.11.0) (2024-03-04)